t085_requirements.txt
# T085 Comprehensive Validation System Requirements
# ML Evaluation Platform Validation Dependencies

# Core async libraries
aiohttp>=3.8.0
asyncio>=3.4.3
asyncpg>=0.28.0

# HTTP and networking
requests>=2.28.0
redis>=4.5.0

# Image processing and test data
Pillow>=9.5.0
numpy>=1.24.0

# YAML parsing for docker-compose.yml
PyYAML>=6.0

# JSON handling (built-in, but specified for clarity)
# json (built-in)

# System and process management
# subprocess (built-in)
# pathlib (built-in)

# Date and time handling
# datetime (built-in)

# Data structures
# dataclasses (built-in)
# typing (built-in)

# Logging
# logging (built-in)

# Optional: Development and debugging tools
pytest>=7.0.0  # For running the validation as a test
pytest-asyncio>=0.21.0  # For async test support